LPS-54176 Rename package
