+++ portal-test-internal/src/com/liferay/portlet/subscriptions/test/BaseSubscriptionLocalizedContentTestCase.java
@@ -19,13 +19,13 @@ import com.dumbster.smtp.MailMessage;
 import com.liferay.portal.kernel.settings.ModifiableSettings;
 import com.liferay.portal.kernel.settings.Settings;
 import com.liferay.portal.kernel.settings.SettingsFactoryUtil;
-import com.liferay.portal.kernel.test.util.MailServiceTestUtil;
 import com.liferay.portal.kernel.util.LocaleThreadLocal;
 import com.liferay.portal.kernel.util.LocaleUtil;
 import com.liferay.portal.kernel.util.LocalizationUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.Layout;
 import com.liferay.portal.util.test.LayoutTestUtil;
+import com.liferay.portal.util.test.MailServiceTestUtil;
 
 import java.util.HashMap;
 import java.util.List;
